Add functionality to safely show jei usage/recipe
